Add RVException class.
This is is meant to represent exceptions as described in the Risc-V spec, as opposed to in Python. In this case, the terms are strongly related, as we have a (hardware) exception represented by a (Python) exception. Add a couple of enums to represent exception causes, and refactor IllegalInstruction as an RVException so that when the trap mechanism is built, that will just work.